add fallback method

+++ book-service/src/main/java/uz/pdp/book_service/controller/BookController.java
@@ -1,5 +1,6 @@
 package uz.pdp.book_service.controller;
 
+import io.github.resilience4j.circuitbreaker.annotation.CircuitBreaker;
 import jakarta.validation.Valid;
 import lombok.RequiredArgsConstructor;
 import org.springframework.web.bind.annotation.*;
@@ -39,7 +40,7 @@ public class BookController {
     }
 
     @DeleteMapping("/studentId/{studentId}")
-    public Response<?> deleteBookByStudentId(@PathVariable Long studentId) {
+    public Response<?> deleteBookByStudentId(@PathVariable Integer studentId) {
         return bookService.deleteBookByStudentId(studentId);
     }
 }
